wallet: move deprecated tests to deprecated_test.go
