+++ wallet/wallet_test.go
@@ -2,22 +2,23 @@ package wallet
 
 import (
 	"encoding/hex"
+	"errors"
 	"fmt"
-	"math"
-	"strings"
-	"sync"
-	"sync/atomic"
 	"testing"
 	"time"
 
 	"github.com/btcsuite/btcd/btcutil"
 	"github.com/btcsuite/btcd/chaincfg/chainhash"
 	"github.com/btcsuite/btcd/wire"
-	"github.com/btcsuite/btcwallet/waddrmgr"
-	"github.com/btcsuite/btcwallet/walletdb"
 	"github.com/btcsuite/btcwallet/wtxmgr"
-	"github.com/stretchr/testify/require"
-	"golang.org/x/sync/errgroup"
+)
+
+var (
+	// errBlockNotFound is an error returned when a block is not found.
+	errBlockNotFound = errors.New("block not found")
+
+	// errHeaderNotFound is an error returned when a header is not found.
+	errHeaderNotFound = errors.New("header not found")
 )
 
 var (
@@ -34,6 +35,84 @@ var (
 		Time: time.Now(),
 	}
 )
+
+// mockChainConn is a mock in-memory implementation of the chainConn interface
+// that will be used for the birthday block sanity check tests. The struct is
+// capable of being backed by a chain in order to reproduce real-world
+// scenarios.
+type mockChainConn struct {
+	chainTip    uint32
+	blockHashes map[uint32]chainhash.Hash
+	blocks      map[chainhash.Hash]*wire.MsgBlock
+}
+
+var _ chainConn = (*mockChainConn)(nil)
+
+// createMockChainConn creates a new mock chain connection backed by a chain
+// with N blocks. Each block has a timestamp that is exactly blockInterval after
+// the previous block's timestamp.
+func createMockChainConn(genesis *wire.MsgBlock, n uint32,
+	blockInterval time.Duration) *mockChainConn {
+
+	c := &mockChainConn{
+		chainTip:    n,
+		blockHashes: make(map[uint32]chainhash.Hash),
+		blocks:      make(map[chainhash.Hash]*wire.MsgBlock),
+	}
+
+	genesisHash := genesis.BlockHash()
+	c.blockHashes[0] = genesisHash
+	c.blocks[genesisHash] = genesis
+
+	for i := uint32(1); i <= n; i++ {
+		prevTimestamp := c.blocks[c.blockHashes[i-1]].Header.Timestamp
+		block := &wire.MsgBlock{
+			Header: wire.BlockHeader{
+				Timestamp: prevTimestamp.Add(blockInterval),
+			},
+		}
+
+		blockHash := block.BlockHash()
+		c.blockHashes[i] = blockHash
+		c.blocks[blockHash] = block
+	}
+
+	return c
+}
+
+// GetBestBlock returns the hash and height of the best block known to the
+// backend.
+func (c *mockChainConn) GetBestBlock() (*chainhash.Hash, int32, error) {
+	bestHash, ok := c.blockHashes[c.chainTip]
+	if !ok {
+		return nil, 0, fmt.Errorf("%w: height %d",
+			errBlockNotFound, c.chainTip)
+	}
+
+	return &bestHash, int32(c.chainTip), nil
+}
+
+// GetBlockHash returns the hash of the block with the given height.
+func (c *mockChainConn) GetBlockHash(height int64) (*chainhash.Hash, error) {
+	hash, ok := c.blockHashes[uint32(height)]
+	if !ok {
+		return nil, fmt.Errorf("%w: height %d", errBlockNotFound, height)
+	}
+
+	return &hash, nil
+}
+
+// GetBlockHeader returns the header for the block with the given hash.
+func (c *mockChainConn) GetBlockHeader(
+	hash *chainhash.Hash) (*wire.BlockHeader, error) {
+
+	block, ok := c.blocks[*hash]
+	if !ok {
+		return nil, fmt.Errorf("%w: hash %v", errHeaderNotFound, hash)
+	}
+
+	return &block.Header, nil
+}
 
 // TestLocateBirthdayBlock ensures we can properly map a block in the chain to a
 // timestamp.
@@ -112,544 +191,5 @@ func TestLocateBirthdayBlock(t *testing.T) {
 		if !success {
 			break
 		}
-	}
-}
-
-// TestLabelTransaction tests labelling of transactions with invalid labels,
-// and failure to label a transaction when it already has a label.
-func TestLabelTransaction(t *testing.T) {
-	t.Parallel()
-
-	tests := []struct {
-		name string
-
-		// Whether the transaction should be known to the wallet.
-		txKnown bool
-
-		// Whether the test should write an existing label to disk.
-		existingLabel bool
-
-		// The overwrite parameter to call label transaction with.
-		overwrite bool
-
-		// The error we expect to be returned.
-		expectedErr error
-	}{
-		{
-			name:          "existing label, not overwrite",
-			txKnown:       true,
-			existingLabel: true,
-			overwrite:     false,
-			expectedErr:   ErrTxLabelExists,
-		},
-		{
-			name:          "existing label, overwritten",
-			txKnown:       true,
-			existingLabel: true,
-			overwrite:     true,
-			expectedErr:   nil,
-		},
-		{
-			name:          "no prexisting label, ok",
-			txKnown:       true,
-			existingLabel: false,
-			overwrite:     false,
-			expectedErr:   nil,
-		},
-		{
-			name:          "transaction unknown",
-			txKnown:       false,
-			existingLabel: false,
-			overwrite:     false,
-			expectedErr:   ErrUnknownTransaction,
-		},
-	}
-
-	for _, test := range tests {
-		test := test
-
-		t.Run(test.name, func(t *testing.T) {
-			w := testWallet(t)
-
-			// If the transaction should be known to the store, we
-			// write txdetail to disk.
-			if test.txKnown {
-				rec, err := wtxmgr.NewTxRecord(
-					TstSerializedTx, time.Now(),
-				)
-				if err != nil {
-					t.Fatal(err)
-				}
-
-				err = walletdb.Update(w.db,
-					func(tx walletdb.ReadWriteTx) error {
-
-						ns := tx.ReadWriteBucket(
-							wtxmgrNamespaceKey,
-						)
-
-						return w.txStore.InsertTx(
-							ns, rec, nil,
-						)
-					})
-				if err != nil {
-					t.Fatalf("could not insert tx: %v", err)
-				}
-			}
-
-			// If we want to setup an existing label for the purpose
-			// of the test, write one to disk.
-			if test.existingLabel {
-				err := w.LabelTransaction(
-					*TstTxHash, "existing label", false,
-				)
-				if err != nil {
-					t.Fatalf("could not write label: %v",
-						err)
-				}
-			}
-
-			newLabel := "new label"
-			err := w.LabelTransaction(
-				*TstTxHash, newLabel, test.overwrite,
-			)
-			if err != test.expectedErr {
-				t.Fatalf("expected: %v, got: %v",
-					test.expectedErr, err)
-			}
-		})
-	}
-}
-
-// TestGetTransaction tests if we can fetch a mined, an existing
-// and a non-existing transaction from the wallet like we expect.
-func TestGetTransaction(t *testing.T) {
-	t.Parallel()
-	rec, err := wtxmgr.NewTxRecord(TstSerializedTx, time.Now())
-	require.NoError(t, err)
-
-	tests := []struct {
-		name string
-
-		// Transaction id.
-		txid chainhash.Hash
-
-		// Expected height.
-		expectedHeight int32
-
-		// Store function.
-		f func(wtxmgr.TxStore,
-			walletdb.ReadWriteBucket) (wtxmgr.TxStore, error)
-
-		// The error we expect to be returned.
-		expectedErr error
-	}{
-		{
-			name:           "existing unmined transaction",
-			txid:           *TstTxHash,
-			expectedHeight: -1,
-			// We write txdetail for the tx to disk.
-			f: func(s wtxmgr.TxStore, ns walletdb.ReadWriteBucket) (
-				wtxmgr.TxStore, error) {
-
-				err = s.InsertTx(ns, rec, nil)
-				return s, err
-			},
-			expectedErr: nil,
-		},
-		{
-			name: "existing mined transaction",
-			txid: *TstTxHash,
-			// We write txdetail for the tx to disk.
-			f: func(s wtxmgr.TxStore, ns walletdb.ReadWriteBucket) (
-				wtxmgr.TxStore, error) {
-
-				err = s.InsertTx(ns, rec, TstMinedSignedTxBlockDetails)
-				return s, err
-			},
-			expectedHeight: TstMinedTxBlockHeight,
-			expectedErr:    nil,
-		},
-		{
-			name: "non-existing transaction",
-			txid: *TstTxHash,
-			// Write no txdetail to disk.
-			f: func(s wtxmgr.TxStore, _ walletdb.ReadWriteBucket) (
-				wtxmgr.TxStore, error) {
-
-				return s, nil
-			},
-			expectedErr: ErrNoTx,
-		},
-	}
-	for _, test := range tests {
-		test := test
-
-		t.Run(test.name, func(t *testing.T) {
-			w := testWallet(t)
-
-			err := walletdb.Update(w.db, func(rw walletdb.ReadWriteTx) error {
-				ns := rw.ReadWriteBucket(wtxmgrNamespaceKey)
-				_, err := test.f(w.txStore, ns)
-				return err
-			})
-			require.NoError(t, err)
-			tx, err := w.GetTransaction(test.txid)
-			require.ErrorIs(t, err, test.expectedErr)
-
-			// Discontinue if no transaction were found.
-			if err != nil {
-				return
-			}
-
-			// Check if we get the expected hash.
-			require.Equal(t, &test.txid, tx.Summary.Hash)
-
-			// Check the block height.
-			require.Equal(t, test.expectedHeight, tx.Height)
-		})
-	}
-}
-
-// TestGetTransactionConfirmations tests that GetTransaction correctly
-// calculates confirmations for both confirmed and unconfirmed transactions.
-// This is a regression test for a bug where confirmations were set to the
-// block height instead of being calculated as currentHeight - blockHeight + 1.
-//
-// The bug had several negative impacts:
-//   - Unconfirmed transactions showed -1 confirmations instead of 0, breaking
-//     zero-conf (accepting transactions before block inclusion)
-//   - Confirmed transactions showed block height instead of actual confirmation
-//     count
-//   - LND and other consumers would make incorrect decisions based on wrong
-//     counts
-func TestGetTransactionConfirmations(t *testing.T) {
-	t.Parallel()
-
-	rec, err := wtxmgr.NewTxRecord(TstSerializedTx, time.Now())
-	require.NoError(t, err)
-
-	tests := []struct {
-		name string
-
-		// Block height where transaction is mined (-1 for unmined).
-		txBlockHeight int32
-
-		// Current wallet sync height.
-		currentHeight int32
-
-		// Expected confirmations.
-		expectedConfirmations int32
-
-		// Expected height in result.
-		expectedHeight int32
-
-		// Whether to check for non-zero timestamp.
-		expectTimestamp bool
-	}{
-		{
-			name:                  "unconfirmed tx",
-			txBlockHeight:         -1,
-			currentHeight:         100,
-			expectedConfirmations: 0,
-			expectedHeight:        -1,
-			expectTimestamp:       false,
-		},
-		{
-			name:                  "tx with 1 confirmation",
-			txBlockHeight:         100,
-			currentHeight:         100,
-			expectedConfirmations: 1,
-			expectedHeight:        100,
-			expectTimestamp:       true,
-		},
-		{
-			name:                  "tx with 3 confirmations",
-			txBlockHeight:         8,
-			currentHeight:         10,
-			expectedConfirmations: 3,
-			expectedHeight:        8,
-			expectTimestamp:       true,
-		},
-		{
-			name:                  "old tx with many confirmations",
-			txBlockHeight:         1,
-			currentHeight:         1000,
-			expectedConfirmations: 1000,
-			expectedHeight:        1,
-			expectTimestamp:       true,
-		},
-		{
-			name:                  "tx in future block",
-			txBlockHeight:         105,
-			currentHeight:         100,
-			expectedConfirmations: 0,
-			expectedHeight:        105,
-			expectTimestamp:       true,
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			t.Parallel()
-			w := testWallet(t)
-
-			// Set the wallet's synced height.
-			err := walletdb.Update(
-				w.db, func(tx walletdb.ReadWriteTx) error {
-					addrmgrNs := tx.ReadWriteBucket(
-						waddrmgrNamespaceKey,
-					)
-					bs := &waddrmgr.BlockStamp{
-						Height: tt.currentHeight,
-						Hash:   chainhash.Hash{},
-					}
-
-					return w.addrStore.SetSyncedTo(
-						addrmgrNs, bs,
-					)
-				},
-			)
-			require.NoError(t, err)
-
-			// Insert transaction into wallet.
-			err = walletdb.Update(
-				w.db, func(tx walletdb.ReadWriteTx) error {
-					ns := tx.ReadWriteBucket(
-						wtxmgrNamespaceKey,
-					)
-
-					// Create block metadata if transaction
-					// is mined.
-					var blockMeta *wtxmgr.BlockMeta
-					if tt.txBlockHeight != -1 {
-						hash := chainhash.Hash{}
-						height := tt.txBlockHeight
-						block := wtxmgr.Block{
-							Hash:   hash,
-							Height: height,
-						}
-						blockMeta = &wtxmgr.BlockMeta{
-							Block: block,
-							Time:  time.Now(),
-						}
-					}
-
-					return w.txStore.InsertTx(
-						ns, rec, blockMeta,
-					)
-				},
-			)
-			require.NoError(t, err)
-
-			result, err := w.GetTransaction(*TstTxHash)
-			require.NoError(t, err)
-
-			require.Equal(
-				t, tt.expectedConfirmations,
-				result.Confirmations,
-			)
-
-			require.Equal(t, tt.expectedHeight, result.Height)
-
-			if tt.expectTimestamp {
-				require.NotZero(t, result.Timestamp)
-			} else {
-				require.Zero(t, result.Timestamp)
-			}
-
-			// Additional checks for unconfirmed transactions.
-			if tt.txBlockHeight == -1 {
-				require.Nil(t, result.BlockHash)
-				require.Equal(t, int32(0), result.Confirmations)
-			} else {
-				require.NotNil(t, result.BlockHash)
-				// Only expect positive confirmations when tx is
-				// not in a future block.
-				if tt.txBlockHeight <= tt.currentHeight {
-					require.Positive(
-						t, result.Confirmations,
-					)
-				} else {
-					// Confirmed txns in future blocks for
-					// example due to reorg should be
-					// treated as unconfirmed and have 0
-					// confirmations.
-					require.Equal(
-						t, int32(0),
-						result.Confirmations,
-					)
-				}
-			}
-		})
-	}
-}
-
-// TestDuplicateAddressDerivation tests that duplicate addresses are not
-// derived when multiple goroutines are concurrently requesting new addresses.
-func TestDuplicateAddressDerivation(t *testing.T) {
-	w := testWallet(t)
-	var (
-		m           sync.Mutex
-		globalAddrs = make(map[string]btcutil.Address)
-	)
-
-	for o := 0; o < 10; o++ {
-		var eg errgroup.Group
-
-		for n := 0; n < 10; n++ {
-			eg.Go(func() error {
-				addrs := make([]btcutil.Address, 10)
-				for i := 0; i < 10; i++ {
-					addr, err := w.NewAddressDeprecated(
-						0, waddrmgr.KeyScopeBIP0084,
-					)
-					if err != nil {
-						return err
-					}
-
-					addrs[i] = addr
-				}
-
-				m.Lock()
-				defer m.Unlock()
-
-				for idx := range addrs {
-					addrStr := addrs[idx].String()
-					if a, ok := globalAddrs[addrStr]; ok {
-						return fmt.Errorf("duplicate "+
-							"address! already "+
-							"have %v, want to "+
-							"add %v", a, addrs[idx])
-					}
-
-					globalAddrs[addrStr] = addrs[idx]
-				}
-
-				return nil
-			})
-		}
-
-		require.NoError(t, eg.Wait())
-	}
-}
-
-func TestEndRecovery(t *testing.T) {
-	// This is an unconventional unit test, but I'm trying to keep things as
-	// succint as possible so that this test is readable without having to mock
-	// up literally everything.
-	// The unmonitored goroutine we're looking at is pretty deep:
-	// SynchronizeRPC -> handleChainNotifications -> syncWithChain -> recovery
-	// The "deadlock" we're addressing isn't actually a deadlock, but the wallet
-	// will hang on Stop() -> WaitForShutdown() until (*Wallet).recovery gets
-	// every single block, which could be hours depending on hardware and
-	// network factors. The WaitGroup is incremented in SynchronizeRPC, and
-	// WaitForShutdown will not return until handleChainNotifications returns,
-	// which is blocked by a running (*Wallet).recovery loop.
-	// It is noted that the conditions for long recovery are difficult to hit
-	// when using btcwallet with a fresh seed, because it requires an early
-	// birthday to be set or established.
-
-	w := testWallet(t)
-
-	blockHashCalled := make(chan struct{})
-
-	chainClient := &mockChainClient{
-		// Force the loop to iterate about forever.
-		getBestBlockHeight: math.MaxInt32,
-		// Get control of when the loop iterates.
-		getBlockHashFunc: func() (*chainhash.Hash, error) {
-			blockHashCalled <- struct{}{}
-			return &chainhash.Hash{}, nil
-		},
-		// Avoid a panic.
-		getBlockHeader: &wire.BlockHeader{},
-	}
-
-	recoveryDone := make(chan struct{})
-	go func() {
-		defer close(recoveryDone)
-		w.recovery(chainClient, &waddrmgr.BlockStamp{})
-	}()
-
-	getBlockHashCalls := func(expCalls int) {
-		var i int
-		for {
-			select {
-			case <-blockHashCalled:
-				i++
-			case <-time.After(time.Second):
-				t.Fatal("expected BlockHash to be called")
-			}
-			if i == expCalls {
-				break
-			}
-		}
-	}
-
-	// Recovery is running.
-	getBlockHashCalls(3)
-
-	// Closing the quit channel, e.g. Stop() without endRecovery, alone will not
-	// end the recovery loop.
-	w.quitMu.Lock()
-	close(w.quit)
-	w.quitMu.Unlock()
-	// Continues scanning.
-	getBlockHashCalls(3)
-
-	// We're done with this one
-	atomic.StoreUint32(&w.recovering.Load().(*recoverySyncer).quit, 1)
-	select {
-	case <-blockHashCalled:
-	case <-recoveryDone:
-	}
-
-	// Try again.
-	w = testWallet(t)
-
-	// We'll catch the error to make sure we're hitting our desired path. The
-	// WaitGroup isn't required for the test, but does show how it completes
-	// shutdown at a higher level.
-	var err error
-	w.wg.Add(1)
-	recoveryDone = make(chan struct{})
-	go func() {
-		defer w.wg.Done()
-		defer close(recoveryDone)
-		err = w.recovery(chainClient, &waddrmgr.BlockStamp{})
-	}()
-
-	waitedForShutdown := make(chan struct{})
-	go func() {
-		w.WaitForShutdown()
-		close(waitedForShutdown)
-	}()
-
-	// Recovery is running.
-	getBlockHashCalls(3)
-
-	// endRecovery is required to exit the unmonitored goroutine.
-	end := w.endRecovery()
-	select {
-	case <-blockHashCalled:
-	case <-recoveryDone:
-	}
-	<-end
-
-	// testWallet starts a couple of other unrelated goroutines that need to be
-	// killed, so we still need to close the quit channel.
-	w.quitMu.Lock()
-	close(w.quit)
-	w.quitMu.Unlock()
-
-	select {
-	case <-waitedForShutdown:
-	case <-time.After(time.Second):
-		t.Fatal("WaitForShutdown never returned")
-	}
-
-	if !strings.EqualFold(err.Error(), "recovery: forced shutdown") {
-		t.Fatal("wrong error")
 	}
 }
